+++ owlcms/src/main/java/app/owlcms/displays/scoreboard/ScoreMultiRanks.java
@@ -204,7 +204,7 @@ public class ScoreMultiRanks extends PolymerTemplate<ScoreMultiRanks.ScoreboardM
         if (event instanceof UIEvent.BreakStarted) {
             UIEvent.BreakStarted e = (UIEvent.BreakStarted) event;
             ceremonyGroup = e.getCeremonyGroup();
-            logger.warn("break event = {} {} {}", e.getBreakType(), e.getTrace(), ceremonyGroup);
+            //logger.trace("break event = {} {} {}", e.getBreakType(), e.getTrace(), ceremonyGroup);
         }
         OwlcmsSession.withFop(fop -> UIEventProcessor.uiAccess(this, uiEventBus, () -> {
             ScoreboardModel model = getModel();